Review round 2: guards in their own file, no migration, re-verify repairs
The tool-loop guards move to lib/runtime/toolLoopGuards.ts, and the
streak keeps SHA-256 digests of the arguments and result instead of the
text, so a large or binary result never lands on the serialized frame.
The saved-policy migration is removed; the placeholder is renamed
AGENCY_INSTALL_DIR_PLACEHOLDER / expandAgencyInstallDir; the policy
label is back to its short form. gradeRun's flag is hasAgencyTests.
codeAgent re-runs the verifier after a successful repair on both
unattended paths and returns those findings, headed "Re-check found
these problems" when errors remain. Also drops a duplicate import that
failed the typecheck job.
