+++ packages/agency-lang/lib/agents/agency-agent/lib/policy.agency
@@ -3,7 +3,6 @@ import {
   cliPolicyHandler,
   parsePolicyFile,
   setPolicy,
-  migrateCatchAllReads,
   minimalAutoApprovePolicy,
   recommendedAutoApprovePolicy,
   builtinPolicy,
@@ -42,7 +41,7 @@ export def givePolicyChoice() {
   },
     {
     key: "recommended",
-    label: "recommended default policy, allow reading files in the current directory (and agency's own docs and skills) and browsing the web, no writes"
+    label: "recommended default policy, allow reading files and browsing the web, no writes"
   },
   ]
 
@@ -79,18 +78,7 @@ export def getPolicyForAgent(
         process.exit(1)
       }
     } else {
-      // A saved copy of an older recommended policy approved every read
-      // anywhere. Scope those rules the way the built-in now does, once,
-      // and say so; a read rule the user wrote stays as written.
-      const updated = migrateCatchAllReads(policy.value)
-      if (updated.migrated.length > 0) {
-        print(
-          color.yellow(
-            "Updated ${POLICY_PATH}: ${updated.migrated.join(", ")} now approve reads only in the current directory and the agency install (they approved reads anywhere). Edit the file to change this.",
-          ),
-        )
-      }
-      setPolicy(POLICY_PATH, updated.policy)
+      setPolicy(POLICY_PATH, policy.value)
     }
   } else {
     const builtin = builtinPolicy(_policyArg, cwd())
