+++ packages/agency-lang/lib/runtime/prompt.ts
@@ -34,6 +34,18 @@ import { PromptBailout, PromptRunner } from "./promptRunner.js";
 import { findIntrinsic, partitionIntrinsicCalls, runIntrinsicCall } from "./intrinsicTools.js";
 import type { RunBatchResult } from "./runBatch.js";
 import { warnOnOversizedToolSchemas } from "./toolSchemaSize.js";
+import {
+  DEFAULT_MAX_REPEATED_TOOL_CALLS,
+  freshRepeatStreak,
+  markupArgument,
+  markupArgumentMessage,
+  noteRepeat,
+  repeatedCallMessage,
+  repeatKey,
+  repeatsBefore,
+  resetRepeat,
+  type RepeatStreak,
+} from "./toolLoopGuards.js";
 import { GuardTripRetry, raiseGuardTripsUntilClear } from "./guardTripInterrupt.js";
 import { failure, isFailure, isSuccess, markDestructiveWork } from "./result.js";
 import type { SourceLocationOpts } from "./state/checkpointStore.js";
@@ -173,105 +185,6 @@ function toolErrorMessage(error: any): string {
  *  breaker against retry spirals), or immediately on the destructive tier. */
 const MAX_TOOL_FAILURES = 5;
 
-/** A call is refused once the same tool, with the same arguments, has
- *  returned the same result this many times in a row, with no other call in
- *  between. That is the loop signature: nothing changed, yet the model asks
- *  again (a writer once made 45 identical, successful typecheck calls). The
- *  refusal also restarts the count, so the call is interrupted every N
- *  repeats rather than banned: a status poll that really is waiting on the
- *  world gets to run again after the model says so. `0` disables. */
-const DEFAULT_MAX_REPEATED_TOOL_CALLS = 3;
-
-/** The name of an optional string argument whose value is the model's own
- *  tool-call markup, or null. Claude sometimes emits the closing tag of its
- *  call syntax where an optional string it meant to leave empty belongs,
- *  and the next parameter's text leaks in:
- *  `stdin: "</antml name=\"stdin\">\n<parameter name=\"allowedExecutables\">[]"`.
- *  Running that call wastes a round at best (`grep` rejects the regex
- *  flags) and at worst executes with garbage. Deliberately narrow: only a
- *  parameter with a default, and only a value that IS the closing tag
- *  (named for this argument, or followed by leaked parameter markup). A
- *  required parameter, or a string that merely contains such text, is data
- *  a transcript or XML tool may legitimately be given. */
-function markupArgument(
-  args: Record<string, unknown>,
-  params: readonly FuncParam[],
-): string | null {
-  for (const param of params) {
-    if (!param.hasDefault) continue;
-    const value = args[param.name];
-    if (typeof value !== "string") continue;
-    const tag = /^<\/antml[^>]*>/.exec(value);
-    if (tag === null) continue;
-    const rest = value.slice(tag[0].length);
-    if (tag[0].includes(param.name) || rest.trimStart().startsWith("<parameter")) {
-      return param.name;
-    }
-  }
-  return null;
-}
-
-/** JSON with object keys sorted at every level, so two calls that pass the
- *  same arguments in a different order get the same key. */
-function canonicalJson(value: unknown): string {
-  if (Array.isArray(value)) return `[${value.map(canonicalJson).join(",")}]`;
-  if (value !== null && typeof value === "object") {
-    const entries = Object.entries(value as Record<string, unknown>)
-      .sort(([a], [b]) => (a < b ? -1 : a > b ? 1 : 0))
-      .map(([k, v]) => `${JSON.stringify(k)}:${canonicalJson(v)}`);
-    return `{${entries.join(",")}}`;
-  }
-  return JSON.stringify(value) ?? "null";
-}
-
-function repeatKey(toolName: string, args: Record<string, unknown>): string {
-  return `${toolName}\u0000${canonicalJson(args)}`;
-}
-
-/** The current run of identical calls: one record, because only calls in a
- *  row count. Any other call, a different result, or a refusal resets it. */
-type RepeatStreak = { key: string; result: string; count: number };
-
-const NO_STREAK: RepeatStreak = { key: "", result: "", count: 0 };
-
-/** Record one completed call and return how many times in a row this exact
- *  call has now produced this exact result. */
-function noteRepeat(streak: RepeatStreak, key: string, result: string): number {
-  if (streak.key === key && streak.result === result) {
-    streak.count += 1;
-  } else {
-    streak.key = key;
-    streak.result = result;
-    streak.count = 1;
-  }
-  return streak.count;
-}
-
-/** How many identical runs in a row precede a call with this key. */
-function repeatsBefore(streak: RepeatStreak, key: string): number {
-  return streak.key === key ? streak.count : 0;
-}
-
-function resetRepeat(streak: RepeatStreak): void {
-  Object.assign(streak, NO_STREAK);
-}
-
-function repeatedCallMessage(toolName: string, count: number): string {
-  return (
-    `Error: this is call ${count + 1} to ${toolName} with exactly these arguments, and ` +
-    `the previous ${count} all returned the same result. It was not run. Say what you ` +
-    `expected to change, then either call it with different arguments or continue without it.`
-  );
-}
-
-function markupArgumentMessage(toolName: string, argument: string): string {
-  return (
-    `Error: the \`${argument}\` argument contains tool-call markup (\`</antml...\`), which ` +
-    `means it was meant to be empty. This call was not run. Call ${toolName} again and ` +
-    `leave \`${argument}\` out.`
-  );
-}
-
 type FailureTier = "destructive" | "neverStarted" | "idempotent" | "neutral";
 
 /** Classify a tool failure. Most-specific fact wins: a started destructive
@@ -346,12 +259,6 @@ export const _internal = {
   failureTier,
   TIER_SUFFIX,
   MAX_TOOL_FAILURES,
-  DEFAULT_MAX_REPEATED_TOOL_CALLS,
-  markupArgument,
-  repeatKey,
-  noteRepeat,
-  repeatsBefore,
-  resetRepeat,
   armCallTimeout,
   runWithRetry,
   dropNullDefaultedArgs,
@@ -741,7 +648,7 @@ export async function runPrompt(args: {
     self.__initialized = true;
     self.removedTools = args.removedTools || [];
     self.toolErrorCounts = {};
-    self.repeatStreak = { ...NO_STREAK };
+    self.repeatStreak = freshRepeatStreak();
     self.toolCallRound = 0;
     self.validationAttempt = 0;
     self.messagesJSON = null;
